% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{longbc}
\alias{longbc}
\title{Biclustering Multivariate Longitudinal Data}
\usage{
longbc(
  dat,
  lambdas,
  V = NULL,
  tol.abs = 0.01,
  tol.rel = 0.001,
  niter = 50,
  rho = 1,
  tausd = 1/4,
  XtXrhoDeltai = NULL,
  verbose = TRUE,
  trace = TRUE
)
}
\arguments{
\item{dat}{The input data.frame with columns "subj","feat","t", and "y" for the subject, feature, time, and response variables, respectively.}

\item{V}{The plug-in estimator for the covariance matrix of the random intercepts. If NULL, a compound symmetric correlation structure will be estimated from the data.}

\item{tol.abs}{The absolute convergence tolerance}

\item{tol.rel}{The relative convergence tolerance}

\item{niter}{The maximum number of iterations}

\item{rho}{ADMM parameter}

\item{tausd}{Thresholding parameter}

\item{XtXrhoDeltai}{Precomputed sparse matrix}

\item{verbose}{Flag for logging}

\item{trace}{Flag for objective function tracking}

\item{lambdas}{A vector of regularization parameter values}
}
\description{
Implementation of the model proposed in "Biclustering Multivariate Longitudinal Data with Application to a Diffusion Tensor Imaging Study".
}
\author{
Caleb Weaver
}
