% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated data set for demonstration}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{subj}{Subject index}
 \item{feat}{Feature index}
 \item{time}{Time point}
 \item{y}{Response value}
}
}
\usage{
data(simdat)
}
\description{
Simulated data set for demonstration
}
\examples{

data(simdat)
head(simdat)

}
\keyword{datasets}
