\name{longbc-package}
\alias{longbc-package}
\alias{longbc}
\docType{package}
\title{
\packageTitle{longbc}
}
\description{
\packageDescription{longbc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{longbc}
\packageIndices{longbc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{longbc}

Maintainer: \packageMaintainer{longbc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
